+++ setup.cfg
@@ -1,5 +1,6 @@
 [flake8]
 max-line-length = 100
+exclude = __init__.py
 
 [pydocstyle]
 convention = google
